requirements.txt
fastapi
uvicorn[standard]
gunicorn
jinja2
python-multipart
anthropic
sqlalchemy
python-dotenv
